+++ server/test/rss/e2e/admin-accept.e2e-spec.ts
@@ -19,9 +19,9 @@ import { AdminFixture } from '@test/config/common/fixture/admin.fixture';
 import { RssFixture } from '@test/config/common/fixture/rss.fixture';
 import { testApp } from '@test/config/e2e/env/jest.setup';
 
-const URL = '/api/rss/accept';
+const URL = (id: number) => `/api/admins/rss/${id}/acceptances`;
 
-describe(`POST ${URL}/{rssId} E2E Test`, () => {
+describe(`POST /api/admins/rss/{rssId}/acceptances E2E Test`, () => {
   let agent: TestAgent;
   let rssRepository: RssRepository;
   let rssAcceptRepository: RssAcceptRepository;
@@ -55,7 +55,7 @@ describe(`POST ${URL}/{rssId} E2E Test`, () => {
 
   it('[401] 관리자 로그인 쿠키가 없을 경우 RSS 승인을 실패한다.', async () => {
     // Http when
-    const response = await agent.post(`${URL}/${Number.MAX_SAFE_INTEGER}`);
+    const response = await agent.post(URL(Number.MAX_SAFE_INTEGER));
 
     // Http then
     const { data } = response.body;
@@ -80,7 +80,7 @@ describe(`POST ${URL}/{rssId} E2E Test`, () => {
   it('[401] 관리자 로그인 쿠키가 만료됐을 경우 RSS 승인을 실패한다.', async () => {
     // Http when
     const response = await agent
-      .post(`${URL}/${Number.MAX_SAFE_INTEGER}`)
+      .post(URL(Number.MAX_SAFE_INTEGER))
       .set('Cookie', `sessionId=Wrong${sessionKey}`);
 
     // Http then
@@ -106,7 +106,7 @@ describe(`POST ${URL}/{rssId} E2E Test`, () => {
   it('[404] 대기 목록에 없는 RSS를 승인할 경우 RSS 승인을 실패한다.', async () => {
     // Http when
     const response = await agent
-      .post(`${URL}/${Number.MAX_SAFE_INTEGER}`)
+      .post(URL(Number.MAX_SAFE_INTEGER))
       .set('Cookie', `sessionId=${sessionKey}`);
 
     // Http then
@@ -135,7 +135,7 @@ describe(`POST ${URL}/{rssId} E2E Test`, () => {
 
     // Http when
     const response = await agent
-      .post(`${URL}/${rss.id}`)
+      .post(URL(rss.id))
       .set('Cookie', `sessionId=${sessionKey}`);
 
     // Http then
@@ -166,7 +166,7 @@ describe(`POST ${URL}/{rssId} E2E Test`, () => {
 
     // Http when
     const response = await agent
-      .post(`${URL}/${rss.id}`)
+      .post(URL(rss.id))
       .set('Cookie', `sessionId=${sessionKey}`);
 
     // Http then
